+++ saleor/csv/tests/export/products_data/utils.py
@@ -44,6 +44,10 @@ def get_attribute_value(assigned_attribute):
             value += f" {attribute.unit}"
     elif attribute.input_type == AttributeInputType.RICH_TEXT:
         value = clean_editor_js(value_instance.rich_text, to_string=True)
+    elif attribute.input_type == AttributeInputType.SWATCH:
+        value = (
+            value_instance.file_url if value_instance.file_url else value_instance.value
+        )
     elif attribute.input_type == AttributeInputType.BOOLEAN:
         value = str(value_instance.boolean)
     elif attribute.input_type == AttributeInputType.DATE:
@@ -51,7 +55,7 @@ def get_attribute_value(assigned_attribute):
     elif attribute.input_type == AttributeInputType.DATE_TIME:
         value = str(value_instance.date_time)
     else:
-        value = value_instance.slug
+        value = value_instance.name or value_instance.slug
     return value