+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Extend app by `AppExtension` - #7701 by @korycins
 - Add ExternalNotificationTrigger mutation - #7821 by @mstrumeck
 - Add Click&Collect feature - #7673 by @kuchichan
+- Introduce swatch attributes - #7261 by @IKarbowiak
 - Introduce gift card feature - #7827 by @IKarbowiak, @tomaszszymanski129
 
 # 3.0.0 [Unreleased]